Change to int

+++ src/java/syafa/java/hackerrank/TenDaysOfStatistics/Day7SpearmansRankCorrelationCoefficient.java
@@ -11,7 +11,7 @@ public class Day7SpearmansRankCorrelationCoefficient
         double[] firstSet = Stream.of(input.nextLine().split("\s")).mapToDouble(Double::parseDouble).toArray();
         double[] secondSet = Stream.of(input.nextLine().split("\s")).mapToDouble(Double::parseDouble).toArray();
         input.close();
-        double rankDifference = 0d;
+        int rankDifference = 0;
         HashMap<Double,Integer> firstSetMap = new HashMap<Double,Integer>();
         HashMap<Double,Integer> secondSetMap = new HashMap<Double,Integer>();
         for (int i = 0; i < (firstSet.length + secondSet.length) / 2; i++)
